Integrated MSBASIC
Integrated MSBASIC into the G80 System with Doug's Monitor and My DOS for CH376 module.  Note that there seems to be a problem with listing basic programs for line number under 255 it prints 0 for line number.  I saw this months ago and cannot remember what fixed it.  I think it might be the pasmo compiler - I am using zasm on my mac and it all works well... Also the SCREEN and COLOR commands are broken...