Official V1.0
Fixed the MS Basic line number less than 255 issue. was a bug in PASMO assembler - changed code to CMP -24-8 to CMP 0E0H.  The SCREEN and COLOR are not working as well.  Moved the Monitor Jump table to 0100H instead of 0D6 so there is room no matter which languages you select.